RFT: Consolidate L1 tests.
